Make login required to access endpoints

+++ app/products/routers.py
@@ -1,5 +1,7 @@
+from auth.routers import get_current_user
+from auth.schemas import User
 from db import get_db
-from fastapi import APIRouter, Depends, status
+from fastapi import APIRouter, Depends, HTTPException, status
 from products import cruds, schemas
 from sqlalchemy.orm import Session
 
@@ -17,15 +19,35 @@ async def get_product(product_id: int, db: Session = Depends(get_db)):
 
 
 @router.delete("/{product_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_product(product_id: int, db: Session = Depends(get_db)):
-    cruds.delete_product(db, product_id)
+async def delete_product(
+    product_id: int, current_user: User = Depends(get_current_user), db: Session = Depends(get_db)
+):
+    if current_user.is_admin:
+        cruds.delete_product(db, product_id)
+    else:
+        raise HTTPException(status_code=403, detail="Not an admin user!")
 
 
 @router.post("", response_model=schemas.ProductCreate)
-async def create_product(product: schemas.ProductCreate, db: Session = Depends(get_db)):
-    return cruds.create_product(db, product)
+async def create_product(
+    product: schemas.ProductCreate,
+    current_user: User = Depends(get_current_user),
+    db: Session = Depends(get_db),
+):
+    if current_user.is_admin:
+        return cruds.create_product(db, product)
+    else:
+        raise HTTPException(status_code=403, detail="Not an admin user!")
 
 
 @router.put("/{product_id}")
-async def update_product(product_id: int, product: schemas.ProductUpdate, db: Session = Depends(get_db)):
-    return cruds.update_product(db, product_id, product)
+async def update_product(
+    product_id: int,
+    product: schemas.ProductUpdate,
+    current_user: User = Depends(get_current_user),
+    db: Session = Depends(get_db),
+):
+    if current_user.is_admin:
+        return cruds.update_product(db, product_id, product)
+    else:
+        raise HTTPException(status_code=403, detail="Not an admin user!")
